combined achievement and achievementDetail, added type, refactoring
