+++ src/test/java/juja/microservices/gamification/integration/AchievementValidationJsonIntegrationTest.java
@@ -1,7 +1,6 @@
-package juja.microservices.gamification.achievement;
+package juja.microservices.gamification.integration;
 
 import com.lordofthejars.nosqlunit.annotation.UsingDataSet;
-import juja.microservices.gamification.BaseIntegrationTest;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
